refactor SignUpFormModal to CreateAccountModal
